+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `scope` property for all rule templates.
+
 ## [0.23.1] - 2020-08-25
 
 ### Added