+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.25.0] - 2020-09-01
+
 ### Added
 
 - `ForbiddenPropertyRule` analysis rule template, which flags properties with a specified
@@ -584,7 +586,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.24.0...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.25.0...HEAD
+[0.25.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.23.1...v0.24.0
 [0.23.1]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.23.0...v0.23.1
 [0.23.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.22.1...v0.23.0